Add UserSelectDatePicker for consultations. Add ProposedTermsContent component for consultations. Update DatePicker. Add translations.
